requirements.txt
pytorch-lightning==1.2.10
datasets==1.6.2
transformers==4.5.1
scikit-learn==0.24.2
numpy==1.23.4
packaging==21.3
tabulate>=0.7.7
wandb==0.14.0
hydra-core==1.3.2
hydra_colorlog==1.2.0
dvc==2.50.0
tqdm==4.49.0